include tmg package, update drawing synth and preprocessing
Adds the tar file for tmg package to install from archive. The current
version is no longer available on CRAN and it is required for bfa.

The preprocessing of item response data extracts only the item responses,
violent behavior outcome and the age data for subpopulation results.
Also stores train and test splits based on interview date.

The code for drawing synthetic data draws 3 sets of synthetic data,
one for pruning, one for fitting the tree and one for doing
uncertainty quantification via plotting.